% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar1_matrix.R
\name{ar1_matrix}
\alias{ar1_matrix}
\title{Create a first-order autoregressive covariance matrix}
\usage{
ar1_matrix(n, param_specs = NULL)
}
\arguments{
\item{n}{The dimension of the AR(1) covariance matrix.}

\item{param_specs}{Optional parameter specifications for the covariance
matrix.}
}
\value{
An object of S3 class
\code{torch_openreml.covariance.ar1_matrix.AR1Matrix}.
}
\description{
Creates an \code{AR1Matrix} object representing a first-order autoregressive
covariance matrix. The covariance between observations decays geometrically
with the lag between them: \eqn{V_{ij} = \sigma^2 \rho^{|i-j|}}.
}
\details{
The variance parameter \code{sigma^2 > 0} is enforced by \code{TransformExpPow2} and
the correlation parameter \code{rho} in \verb{(-1, 1)} is enforced by a sigmoid
scaled to \verb{(-1, 1)} by default.
}
\examples{
\dontrun{
mat <- ar1_matrix(4)
mat

free_params <- tensor(c(0.5, 1.0))
mat(free_params)

mat$grad(free_params)

mat$set_param_specs("rho", fixed = TRUE)

mat$free_param_names
mat(tensor1d(0.5))
mat$grad(tensor1d(0.5))
}

}
\seealso{
See the \href{https://torch-openreml.patrickli.org/generated/torch_openreml.covariance.ar1matrix#torch_openreml.covariance.AR1Matrix}{online documentation for \code{AR1Matrix}}
for details about the corresponding Python class.
}
