% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{as.matrix.torch.Tensor}
\alias{as.matrix.torch.Tensor}
\title{Coerce a PyTorch tensor to a matrix}
\usage{
\method{as.matrix}{torch.Tensor}(x, ...)
}
\arguments{
\item{x}{A PyTorch tensor.}

\item{...}{Additional arguments, currently unused.}
}
\value{
An R matrix containing the values of \code{x}.
}
\description{
Converts a PyTorch tensor to an R matrix. The tensor is detached from
the computation graph and moved to the CPU before conversion.
}
