% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{as_matrix_array.torch.Tensor}
\alias{as_matrix_array.torch.Tensor}
\alias{as.matrix.torch.Tensor}
\alias{as.array.torch.Tensor}
\title{Coerce a PyTorch tensor to an R matrix or array}
\usage{
\method{as.matrix}{torch.Tensor}(x, ...)

\method{as.array}{torch.Tensor}(x, ...)
}
\arguments{
\item{x}{A PyTorch tensor.}

\item{...}{Additional arguments, currently unused.}
}
\value{
An R matrix or array containing the values of \code{x}.
}
\description{
Converts a PyTorch tensor to an R matrix or array. The tensor is detached
from the computation graph and moved to the CPU before conversion.
}
