% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment}
\alias{augment}
\title{Create a column-wise augmentation of covariance matrices}
\usage{
augment(...)
}
\arguments{
\item{...}{Two or more operands as positional or named arguments or a single named
list mapping names to operands. Operands may be \code{Matrix} instances or
\code{torch.Tensor} values.}
}
\value{
An object of S3 class
\code{torch_openreml.covariance.augment.Augment}.
}
\description{
Creates an \code{Augment} object representing the column-wise augmentation
(horizontal concatenation) of two or more covariance matrices:
}
\details{
\deqn{V = [A_0 \mid A_1 \mid \ldots]}

Each operand is placed side by side (column-wise). All operands must have
the same number of rows. Each operand may be a trainable \code{Matrix} or a
fixed \code{torch.Tensor}.
}
\examples{
\dontrun{
op <- augment(A = scalar_matrix(3), B = scalar_matrix(3))
free_params <- tensor(c(0.5, 1.0))
op(free_params)

grad_result <- op$grad(free_params)
grad_result[[1]]
grad_result[[2]]

op$free_param_names
op$param_names
op$shape
}

}
\seealso{
See the \href{https://torch-openreml.patrickli.org/generated/torch_openreml.covariance.augment#torch_openreml.covariance.Augment}{online documentation for \code{Augment}}
for details about the corresponding Python class.
}
