% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound_symmetric_matrix.R
\name{compound_symmetric_matrix}
\alias{compound_symmetric_matrix}
\title{Create a compound symmetric covariance matrix}
\usage{
compound_symmetric_matrix(n, param_specs = NULL)
}
\arguments{
\item{n}{The dimension of the compound symmetric covariance matrix.}

\item{param_specs}{Optional parameter specifications for the covariance
matrix.}
}
\value{
An object of S3 class
\code{torch_openreml.covariance.compound_symmetric_matrix.CompoundSymmetricMatrix}.
}
\description{
Creates a \code{CompoundSymmetricMatrix} object representing a compound
symmetric covariance matrix with a shared variance and correlation.
The diagonal entries are equal to the variance, while the off-diagonal
entries are equal to the variance multiplied by the correlation.
}
\details{
The variance parameter is transformed to a positive value using
\code{TransformExpPow2} by default. The correlation parameter is transformed
using a sigmoid scaled to the interval \verb{(-1 / (n - 1), 1)} to ensure
positive definiteness.
}
\examples{
\dontrun{
mat <- compound_symmetric_matrix(3)
mat

free_params <- tensor(c(0.5, 0))
mat(free_params)

mat$grad(free_params)

mat$set_param_specs("rho", fixed = TRUE)

mat$free_param_names
mat(tensor(array(0.5)))
mat$grad(tensor(array(0.5)))
}

}
\seealso{
See the \href{https://torch-openreml.patrickli.org/generated/torch_openreml.covariance.compoundsymmetricmatrix#torch_openreml.covariance.CompoundSymmetricMatrix}{online documentation for \code{CompoundSymmetricMatrix}}
for details about the corresponding Python class.
}
