% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagonal_matrix.R
\name{diagonal_matrix}
\alias{diagonal_matrix}
\title{Create a diagonal covariance matrix}
\usage{
diagonal_matrix(n, param_specs = NULL)
}
\arguments{
\item{n}{The dimension of the diagonal covariance matrix.}

\item{param_specs}{Optional parameter specifications for the diagonal
covariance matrix.}
}
\value{
An object of S3 class
\code{torch_openreml.covariance.diagonal_matrix.DiagonalMatrix}.
}
\description{
Creates a \code{DiagonalMatrix} object representing a diagonal covariance matrix
with one variance parameter per entry. Each diagonal entry is transformed
to a positive variance using \code{TransformExpPow2} by default.
}
\examples{
\dontrun{
mat <- diagonal_matrix(3)
mat

free_params <- tensor(c(0, 0.5, 1))
mat(free_params)

mat$grad(free_params)

mat$set_param_specs("sigma^2_0", fixed = TRUE)

mat$free_param_names
mat(tensor(c(0.5, 1)))
mat$grad(tensor(c(0.5, 1)))
}

}
\seealso{
See the \href{https://torch-openreml.patrickli.org/generated/torch_openreml.covariance.diagonalmatrix#torch_openreml.covariance.DiagonalMatrix}{online documentation for \code{DiagonalMatrix}}
for details about the corresponding Python class.
}
