% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_openreml.R
\name{find_conda}
\alias{find_conda}
\title{Find the Conda executable}
\usage{
find_conda()
}
\value{
A character vector containing the path to the detected Conda
executable, or \code{NULL} if Conda is not installed.
}
\description{
Retrieves the path to the Conda executable detected by
\code{reticulate}.
}
\keyword{internal}
