% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadamard_product.R
\name{hadamard_product}
\alias{hadamard_product}
\title{Create a Hadamard (element-wise) product of two covariance matrices}
\usage{
hadamard_product(...)
}
\arguments{
\item{...}{Exactly two operands as positional or named arguments or a single named
list. The first is \code{A}, the second is \code{B}. Operands may be \code{Matrix}
instances or \code{torch.Tensor} values.}
}
\value{
An object of S3 class
\code{torch_openreml.covariance.hadamard_product.HadamardProduct}.
}
\description{
Creates a \code{HadamardProduct} object representing the element-wise product of
two covariance matrices:
}
\details{
\deqn{V = A \odot B}

Both operands must have the same shape. Either or both may be trainable
\code{Matrix} instances or fixed \code{torch.Tensor} values.
}
\examples{
\dontrun{
n <- 4
op <- hadamard_product(a = equicorrelation_matrix(n), b = tensor(array(5.0)))
free_params <- tensor(array(1.0))
op(free_params)

grad_result <- op$grad(free_params)
grad_result[[1]]
grad_result[[2]]

op$free_param_names
op$param_names
op$shape
}

}
\seealso{
See the \href{https://torch-openreml.patrickli.org/generated/torch_openreml.covariance.hadamardproduct#torch_openreml.covariance.HadamardProduct}{online documentation for \code{HadamardProduct}}
for details about the corresponding Python class.
}
