% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity_matrix.R
\name{identity_matrix}
\alias{identity_matrix}
\title{Create an identity matrix}
\usage{
identity_matrix(n)
}
\arguments{
\item{n}{The dimension of the identity matrix.}
}
\value{
An object of S3 class
\code{torch_openreml.covariance.identity_matrix.IdentityMatrix}.
}
\description{
Creates an \code{IdentityMatrix} object representing an identity matrix of the
specified dimension.
}
\seealso{
See the \href{https://torch-openreml.patrickli.org/generated/torch_openreml.covariance.identitymatrix#torch_openreml.covariance.IdentityMatrix}{online documentation for \code{IdentityMatrix}}
for details about the corresponding Python class.
}
