% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronecker_product.R
\name{kronecker_product}
\alias{kronecker_product}
\title{Create a Kronecker product of two covariance matrices}
\usage{
kronecker_product(...)
}
\arguments{
\item{...}{Exactly two operands as positional or named arguments or a single named
list. The first is \code{A}, the second is \code{B}. Operands may be \code{Matrix}
instances or \code{torch.Tensor} values.}
}
\value{
An object of S3 class
\code{torch_openreml.covariance.kronecker_product.KroneckerProduct}.
}
\description{
Creates a \code{KroneckerProduct} object representing the Kronecker product of
two covariance matrices:
}
\details{
\deqn{V = A \otimes B}

If \code{A} is \verb{m x m} and \code{B} is \verb{n x n}, the result is an \verb{mn x mn} matrix.
Either or both operands may be trainable \code{Matrix} instances or fixed
\code{torch.Tensor} values.
}
\examples{
\dontrun{
op <- kronecker_product(time = ar1_matrix(2), subject = scalar_matrix(2))
params <- tensor(c(1.0, 1.0, 1.0))
op(params)

grad_result <- op$grad(params)
grad_result[[1]]
grad_result[[2]]

op$free_param_names
op$param_names
op$shape
}

}
\seealso{
See the \href{https://torch-openreml.patrickli.org/generated/torch_openreml.covariance.kroneckerproduct#torch_openreml.covariance.KroneckerProduct}{online documentation for \code{KroneckerProduct}}
for details about the corresponding Python class.
}
