% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tensor}
\alias{tensor}
\title{Create a PyTorch tensor}
\usage{
tensor(x, dtype = NULL, device = NULL, ...)
}
\arguments{
\item{x}{Data from which to create the tensor.}

\item{dtype}{Optional PyTorch data type.}

\item{device}{Optional device on which to create the tensor, such as
\code{"cpu"}, \code{"cuda"}, or \code{"mps"}.}

\item{...}{Additional arguments passed to \code{torch.tensor()}.}
}
\value{
A Python PyTorch tensor.
}
\description{
Creates a PyTorch tensor from the supplied data by calling
\href{https://pytorch.org/docs/stable/generated/torch.tensor.html}{\code{torch.tensor()}}.
}
