% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_chain.R
\name{transform_chain}
\alias{transform_chain}
\title{Create a chained transformation}
\usage{
transform_chain(chain)
}
\arguments{
\item{chain}{A \code{Transform} object or a list or tuple of \code{Transform} objects
to compose.}
}
\value{
An object of S3 class
\code{torch_openreml.covariance.transform.TransformChain}.
}
\description{
Creates a \code{TransformChain} object that composes multiple transformations
applied sequentially. The forward transformation applies the transforms in
order, while the inverse transformation applies them in reverse order.
}
\examples{
\dontrun{
t <- transform_chain(list(
  transform_exp(),
  transform_pow(factor = 2.0)
))
x <- tensor(1.0)
t(x)

t <- transform_chain(list(
  transform_exp(),
  transform_pow(factor = 2.0)
))
x <- tensor(4.0)
t$inverse(x)

t <- transform_chain(list(
  transform_exp(),
  transform_pow(factor = 2.0)
))
x <- tensor(1.0)
t$grad(x)
}

}
\seealso{
See the \href{https://torch-openreml.patrickli.org/generated/torch_openreml.covariance.transform.transformchain#torch_openreml.covariance.transform.TransformChain}{online documentation for \code{TransformChain}}
for details about the corresponding Python class.
}
