% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_exp.R
\name{transform_exp}
\alias{transform_exp}
\title{Create an exponential transformation}
\usage{
transform_exp()
}
\value{
An object of S3 class
\code{torch_openreml.covariance.transform.TransformExp}.
}
\description{
Creates a \code{TransformExp} object that applies the natural exponential
function element-wise. The transformation maps from the real numbers to
the positive real numbers.
}
\examples{
\dontrun{
t <- transform_exp()
x <- tensor(c(0.0, 1.0))
t(x)

t <- transform_exp()
x <- tensor(1.0)
t$inverse(x)

t <- transform_exp()
x <- tensor(0.0)
t$grad(x)
}

}
\seealso{
See the \href{https://torch-openreml.patrickli.org/generated/torch_openreml.covariance.transform.transformexp#torch_openreml.covariance.transform.TransformExp}{online documentation for \code{TransformExp}}
for details about the corresponding Python class.
}
