% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_exp10.R
\name{transform_exp10}
\alias{transform_exp10}
\title{Create a base-10 exponential transformation}
\usage{
transform_exp10()
}
\value{
An object of S3 class
\code{torch_openreml.covariance.transform.TransformExp10}.
}
\description{
Creates a \code{TransformExp10} object that applies the base-10 exponential
function element-wise. The transformation maps from the real numbers to
the positive real numbers.
}
\examples{
\dontrun{
t <- transform_exp10()
x <- tensor(c(0.0, 1.0))
t(x)

t <- transform_exp10()
x <- tensor(c(1.0, 10.0))
t$inverse(x)

t <- transform_exp10()
x <- tensor(1.0)
t$grad(x)
}

}
\seealso{
See the \href{https://torch-openreml.patrickli.org/generated/torch_openreml.covariance.transform.transformexp10#torch_openreml.covariance.transform.TransformExp10}{online documentation for \code{TransformExp10}}
for details about the corresponding Python class.
}
