% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_exp2.R
\name{transform_exp2}
\alias{transform_exp2}
\title{Create a base-2 exponential transformation}
\usage{
transform_exp2()
}
\value{
An object of S3 class
\code{torch_openreml.covariance.transform.TransformExp2}.
}
\description{
Creates a \code{TransformExp2} object that applies the base-2 exponential
function element-wise. The transformation maps from the real numbers to
the positive real numbers.
}
\examples{
\dontrun{
t <- transform_exp2()
x <- tensor(c(0.0, 1.0))
t(x)

t <- transform_exp2()
x <- tensor(c(1.0, 2.0))
t$inverse(x)

t <- transform_exp2()
x <- tensor(1.0)
t$grad(x)
}

}
\seealso{
See the \href{https://torch-openreml.patrickli.org/generated/torch_openreml.covariance.transform.transformexp2#torch_openreml.covariance.transform.TransformExp2}{online documentation for \code{TransformExp2}}
for details about the corresponding Python class.
}
