% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_exp_pow2.R
\name{transform_exp_pow2}
\alias{transform_exp_pow2}
\title{Create a scaled exponential transformation}
\usage{
transform_exp_pow2()
}
\value{
An object of S3 class
\code{torch_openreml.covariance.transform.TransformExpPow2}.
}
\description{
Creates a \code{TransformExpPow2} object that applies an exponential
transformation with the exponent scaled by 2. The transformation maps
from the real numbers to the non-negative real numbers.
}
\examples{
\dontrun{
t <- transform_exp_pow2()
x <- tensor(c(0.0, 1.0))
t(x)

t <- transform_exp_pow2()
x <- tensor(1.0)
t$inverse(x)

t <- transform_exp_pow2()
x <- tensor(c(0.0, 1.0))
t$grad(x)
}

}
\seealso{
See the \href{https://torch-openreml.patrickli.org/generated/torch_openreml.covariance.transform.transformexppow2#torch_openreml.covariance.transform.TransformExpPow2}{online documentation for \code{TransformExpPow2}}
for details about the corresponding Python class.
}
