% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_identity.R
\name{transform_identity}
\alias{transform_identity}
\title{Create an identity transformation}
\usage{
transform_identity()
}
\value{
An object of S3 class
\code{torch_openreml.covariance.TransformIdentity}.
}
\description{
Creates a \code{TransformIdentity} object that leaves its input unchanged.
}
\examples{
\dontrun{
t <- transform_identity()
x <- tensor(c(0.0, 1.0, -3.5))
t(x)

t <- transform_identity()
x <- tensor(c(2.0, -1.0))
t$inverse(x)

t <- transform_identity()
x <- tensor(c(0.0, 1.0))
t$grad(x)
}

}
\seealso{
See the \href{https://torch-openreml.patrickli.org/generated/torch_openreml.covariance.transform.TransformIdentity}{online documentation for \code{TransformIdentity}}
for details about the corresponding Python class.
}
