% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_pow.R
\name{transform_pow}
\alias{transform_pow}
\title{Create a power transformation}
\usage{
transform_pow(factor = 2)
}
\arguments{
\item{factor}{The exponent of the power transformation. Defaults to \code{2.0}.}
}
\value{
An object of S3 class
\code{torch_openreml.covariance.transform.TransformPow}.
}
\description{
Creates a \code{TransformPow} object that applies a power transformation with
a configurable exponent.
}
\details{
Negative values raised to a non-integer exponent return \code{torch.nan} with
the same shape as the input. This can occur in both the forward and inverse
transformations.
}
\examples{
\dontrun{
t <- transform_pow(factor = 3.0)
x <- tensor(c(1.0, 2.0, 3.0))
t(x)

t <- transform_pow(factor = 2.0)
x <- tensor(c(1.0, 4.0, 9.0))
t$inverse(x)

t <- transform_pow(factor = 3.0)
x <- tensor(c(2.0, 3.0))
t$grad(x)
}

}
\seealso{
See the \href{https://torch-openreml.patrickli.org/generated/torch_openreml.covariance.transform.transformpow#torch_openreml.covariance.transform.TransformPow}{online documentation for \code{TransformPow}}
for details about the corresponding Python class.
}
