% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_scale_shift.R
\name{transform_scale_shift}
\alias{transform_scale_shift}
\title{Create a scale-shift transformation}
\usage{
transform_scale_shift(a = 0, b = 0)
}
\arguments{
\item{a}{The scale factor.}

\item{b}{The shift offset. Defaults to \code{0.0}.}
}
\value{
An object of S3 class
\code{torch_openreml.covariance.transform.TransformScaleShift}.
}
\description{
Creates a \code{TransformScaleShift} object that applies an affine
transformation with a configurable scale and shift.
}
\examples{
\dontrun{
t <- transform_scale_shift(a = 2.0, b = 1.0)
x <- tensor(c(0.0, 1.0, 2.0))
t(x)

t <- transform_scale_shift(a = 2.0, b = 1.0)
x <- tensor(c(1.0, 3.0, 5.0))
t$inverse(x)

t <- transform_scale_shift(a = 2.0, b = 1.0)
x <- tensor(0.0)
t$grad(x)
}

}
\seealso{
See the \href{https://torch-openreml.patrickli.org/generated/torch_openreml.covariance.transform.transformscaleshift#torch_openreml.covariance.transform.TransformScaleShift}{online documentation for \code{TransformScaleShift}}
for details about the corresponding Python class.
}
