% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_sigmoid.R
\name{transform_sigmoid}
\alias{transform_sigmoid}
\title{Create a sigmoid transformation}
\usage{
transform_sigmoid()
}
\value{
An object of S3 class
\code{torch_openreml.covariance.transform.TransformSigmoid}.
}
\description{
Creates a \code{TransformSigmoid} object that applies the sigmoid function
element-wise. The transformation maps from the real numbers to the open
unit interval.
}
\examples{
\dontrun{
t <- transform_sigmoid()
x <- tensor(c(-2.0, 0.0, 2.0))
t(x)

t <- transform_sigmoid()
x <- tensor(c(0.1, 0.5, 0.9))
t$inverse(x)

t <- transform_sigmoid()
x <- tensor(c(0.0, 1.0))
t$grad(x)
}

}
\seealso{
See the \href{https://torch-openreml.patrickli.org/generated/torch_openreml.covariance.transform.transformsigmoid#torch_openreml.covariance.transform.TransformSigmoid}{online documentation for \code{TransformSigmoid}}
for details about the corresponding Python class.
}
