% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar_matrix.R
\name{scalar_matrix}
\alias{scalar_matrix}
\title{Create a scalar matrix}
\usage{
scalar_matrix(n, param_specs = NULL)
}
\arguments{
\item{n}{The dimension of the scalar matrix.}

\item{param_specs}{Optional parameter specifications for the scalar matrix.}
}
\value{
An object of S3 class
\code{torch_openreml.covariance.scalar_matrix.ScalarMatrix}.
}
\description{
Creates a \code{ScalarMatrix} object representing a scalar matrix of the
specified dimension.
}
\seealso{
See the \href{https://torch-openreml.patrickli.org/generated/torch_openreml.covariance.scalarmatrix#torch_openreml.covariance.ScalarMatrix}{online documentation for \code{ScalarMatrix}}
for details about the corresponding Python class.
}
